backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.5.0
fastapi-cors==0.0.6
pytest==7.4.3
httpx==0.25.2
requests==2.31.0
python-multipart==0.0.6
